requirements.txt

aiogram==2.25.2
python-dotenv
cryptopay==1.1.0
    